+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add TextEditor::AnyState const value.
 - Add TextEditor::add_key_binding() and remove_key_binding().
 - Fix for unicode fonts.
+- Add DisplayExt::set_highlight_data_ext() and text::StyleTableEntryExt (supports underline, strikethrough, background color...etc).
 - Pull FLTK menu fixes.
 - Update FLTK (removes experimental drivers).